+++ supvisors/tests/conftest.py
@@ -208,7 +208,10 @@ def mock_xml_rpc(proxy):
     proxy.supvisors.get_supvisors_state.return_value = {'fsm_statename': 'OPERATION', 'starting_jobs': []}
     proxy.supvisors.get_all_instances_state_modes.return_value = [{'fsm_statename': 'OPERATION'}]
     proxy.supvisors.get_instance_state_modes.return_value = {'fsm_statename': 'OPERATION'}
-    proxy.supvisors.get_master_identifier.return_value = 'server_01'
+    proxy.supvisors.get_master_identifier.return_value = {'identifier': 'server_01:5505',
+                                                          'nick_identifier': 'server_01'}
+    proxy.supvisors.get_core_identifiers.return_value = [{'identifier': 'server_01:5505',
+                                                          'nick_identifier': 'server_01'}]
     proxy.supvisors.get_strategies.return_value = {'auto-fencing': False, 'conciliation': 'USER'}
     proxy.supvisors.get_statistics_status.return_value = {'host_stats': True}
     proxy.supvisors.get_network_info.return_value = {'host': 'localhost'}